+++ .github/CHANGELOG.rst
@@ -9,7 +9,7 @@ Change Log
 
 2.4.4
 -----
-*Release date: TBD*
+*Release date: 13 July 2020*
 
 - Fixed colours associated with adjudicators' scores not showing
 - Showed validation errors when using API with invalid field names
@@ -17,6 +17,7 @@ Change Log
 - Fixed eligibilty API endpoints when a list of participants is not present
 - Fixed speaker category eligibility API endpoint not accepting speakers
 - Allowed updating teams, excluding speakers, through the team detail API endpoint
+- Added date and time formats for Malay
 
 
 2.4.3